Restricted add-ability of certain fields.
Fields now have a limit to the maximum of their type that can be added, e.g. maximum 5 students can be added on the form. Some fields are set to a maximum of 1, which prevents adding any more of the field. Also moved the student info section to a separate variable because it was repeated in both the tutoring and course forms.