Fixed validation for conditional fields, removed redundant add field buttons
Conditional field validation is now properly set up when a conditional field is loaded so that it doesn't immediately appear as an error. Only 1 add field button is shown for fields of the same type (i.e. student names). Added a proxy property to package.json for interacting with the backend (later).