Parent and Student edits should be functional
Added an "edit" button to student profiles which takes the user to a pre-filled form using the student's and parent's information. (Will only work for users pulled from the backend). On submit, if a parent is selected, that parent is `patch`ed, otherwise a new parent post is made. Similarly, for students, if this was an edit form, a patch is made, otherwise a new student is posted.