Fix infinite loading for StudentCourseViewer
Close #192

Previously, if the student had no enrollments, no courses would be requested, and so its status would be null and thus considered to be loading.
Now, only shows a loading/error screen if the enrollments fetch has not complete, avoiding an infinite screen if the student has no enrollments.