Changed the name to Add_profile_spec.js by combining 2 tests
I've combined tests AF and AC into one since they essentially test the notes feature of the app. What the first test will test for is the accounts notes for an instructor. It will test to see if the instructor can create a note, and create another note with a notification. It will also test to see if the notes tab shows the notification correctly. It will then delete the notes and complete the the test. Second is the same but for the courses in registartion.