Moved switch to top of page
Still need to figure out how to theme react admin, but for now I've put the switch on the top of the page. The issue we are having is @neilallavarpu created dynamic buttons that render one ontop of another. The switch was part of the the top form and the imported input was on the bottom.